backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
PyMuPDF==1.23.8
pytesseract==0.3.10
pillow==11.3.0

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3

# Database
sqlalchemy==2.0.23

# LLM integration
ollama==0.1.7
requests==2.31.0

# Text processing
spacy==3.7.2
nltk==3.8.1
textstat==0.7.3

# Image processing
opencv-python==4.8.1.78
easyocr==1.7.0

# Knowledge graph
networkx==3.2.1
rdflib==7.0.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Redis for caching and session management
redis==5.0.1
aioredis==2.0.1

# Enhanced database support
alembic==1.13.1
psycopg2-binary==2.9.9

# Machine learning for topic modeling
scikit-learn==1.3.2

# Performance monitoring and testing
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Docker for secure code execution
docker==6.1.3

# Multi-modal document processing (basic dependencies)
matplotlib==3.8.2
plotly==5.17.0
sympy==1.12
pdfminer.six==20221105

# Voice processing dependencies
SpeechRecognition==3.10.0
pyttsx3==2.90
librosa==0.10.1
soundfile==0.12.1
#pyaudio==0.2.11
#webrtcvad==2.0.10
scipy==1.11.4

# Reference manager integrations
aiohttp==3.9.1
xmltodict==0.13.0
beautifulsoup4==4.12.2

# Note-taking app integrations
PyYAML==6.0.1

# Jupyter notebook support
nbformat==5.9.2
jupyter-client==8.6.0
ipykernel==6.27.1
ipywidgets==8.1.1

# Note: Advanced ML models (transformers, torch, etc.) require additional setup
# For production deployment, install these separately:
# transformers==4.36.0
# torch==2.1.0
# torchvision==0.16.0
# pdf2image==1.16.3
# camelot-py[cv]==0.11.0
# tabula-py==2.9.0

# GraphQL support
strawberry-graphql==0.215.1

# Push notifications
pywebpush==1.14.0

# Additional cryptography and JWT support
cryptography==41.0.7
PyJWT==2.8.0

# WebSocket support
websockets==12.0

# Additional async support
asyncio-mqtt==0.16.1
